requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
openai==1.3.7
python-dotenv==1.0.0
webdriver-manager==4.0.1
linkedin-scraper==2.11.1
google-search-results==2.4.2
tqdm==4.66.1
plotly==5.17.0
dash==2.14.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
schedule==1.2.0
firecrawl==4.3.1
